api/requirements.txt
fastapi
uvicorn
asyncio
python-dotenv
python-multipart
pydantic
pydantic-settings
azure-storage-blob
azure-identity
azure-keyvault-secrets
azure-ai-textanalytics
azure-monitor-opentelemetry
openai>=1.0.0
qdrant-client
tenacity
httpx
python-jose
passlib
bcrypt
pytest
pytest-asyncio
black
flake8
aiohttp